Add arguments and docs
