+++ docs/source/general/predictions.rst
@@ -15,9 +15,8 @@ You can pass in a sample of data (image file path, a string of text, etc) to the
 
 .. code-block:: python
 
-	from flash import Trainer
 	from flash.data.utils import download_data
-	from flash.vision import ImageClassificationData, ImageClassifier
+	from flash.vision import ImageClassifier
 
 
 	# 1. Download the data set
@@ -45,9 +44,83 @@ Predict on a csv file
 
 	# 2. Load the model from a checkpoint
 	model = TabularClassifier.load_from_checkpoint(
-	    "https://flash-weights.s3.amazonaws.com/tabnet_classification_model.pt"
+		"https://flash-weights.s3.amazonaws.com/tabnet_classification_model.pt"
 	)
 
 	# 3. Generate predictions from a csv file! Who would survive?
 	predictions = model.predict("data/titanic/titanic.csv")
 	print(predictions)
+
+
+Serializing predictions
+=======================
+
+To change how predictions are serialized you can attach a :class:`~flash.data.process.Serializer` to your
+:class:`~flash.Task`. For example, you can choose to serialize outputs as probabilities (for more options see the API
+reference below).
+
+
+.. code-block:: python
+
+	from flash.core.classification import Probabilities
+	from flash.data.utils import download_data
+	from flash.vision import ImageClassifier
+
+
+	# 1. Download the data set
+	download_data("https://pl-flash-data.s3.amazonaws.com/hymenoptera_data.zip", 'data/')
+
+	# 2. Load the model from a checkpoint
+	model = ImageClassifier.load_from_checkpoint("https://flash-weights.s3.amazonaws.com/image_classification_model.pt")
+
+	# 3. Attach the Serializer
+	model.serializer = Probabilities()
+
+	# 4. Predict whether the image contains an ant or a bee
+	predictions = model.predict("data/hymenoptera_data/val/bees/65038344_52a45d090d.jpg")
+	print(predictions)
+	# out: [[0.5926494598388672, 0.40735048055648804]]
+
+
+------
+
+
+******************************************
+Classification serializers - API reference
+******************************************
+
+.. _logits:
+
+Logits
+---------------
+
+.. autoclass:: flash.core.classification.Logits
+	:members:
+	:exclude-members: serialize
+
+.. _probabilities:
+
+Probabilities
+-----------------------
+
+.. autoclass:: flash.core.classification.Probabilities
+	:members:
+	:exclude-members: serialize
+
+.. _classes:
+
+Classes
+-----------------------
+
+.. autoclass:: flash.core.classification.Classes
+	:members:
+	:exclude-members: serialize
+
+.. _labels:
+
+Labels
+-----------------------
+
+.. autoclass:: flash.core.classification.Labels
+	:members:
+	:exclude-members: serialize
