+++ flash/text/classification/model.py
@@ -16,7 +16,6 @@ import warnings
 from typing import Callable, Mapping, Optional, Sequence, Type, Union
 
 import torch
-from torchmetrics import Accuracy
 from transformers import BertForSequenceClassification
 from transformers.modeling_outputs import SequenceClassifierOutput
 
@@ -33,6 +32,8 @@ class TextClassifier(ClassificationTask):
         optimizer: Optimizer to use for training, defaults to `torch.optim.Adam`.
         metrics: Metrics to compute for training and evaluation.
         learning_rate: Learning rate to use for training, defaults to `1e-3`
+        multi_label: Whether the targets are multi-label or not.
+        serializer: The :class:`~flash.data.process.Serializer` to use when serializing prediction outputs.
     """
 
     def __init__(
@@ -40,9 +41,10 @@ class TextClassifier(ClassificationTask):
         num_classes: int,
         backbone: str = "prajjwal1/bert-tiny",
         optimizer: Type[torch.optim.Optimizer] = torch.optim.Adam,
-        metrics: Union[Callable, Mapping, Sequence, None] = [Accuracy()],
+        metrics: Union[Callable, Mapping, Sequence, None] = None,
         learning_rate: float = 1e-3,
-        serializer: Optional[Serializer] = None,
+        multi_label: bool = False,
+        serializer: Optional[Union[Serializer, Mapping[str, Serializer]]] = None,
     ):
         self.save_hyperparameters()
 
@@ -58,6 +60,7 @@ class TextClassifier(ClassificationTask):
             optimizer=optimizer,
             metrics=metrics,
             learning_rate=learning_rate,
+            multi_label=multi_label,
             serializer=serializer,
         )
         self.model = BertForSequenceClassification.from_pretrained(backbone, num_labels=num_classes)
